# go_api/docs/swagger.yaml
definitions:
...
  dto.CreateAcademicCalendarRequest:
    properties:
      academic_day_of_week:
        type: integer
      academic_year:
        type: string
      calendar_date:
        type: string
      description:
        type: string
      semester_type:
        type: string
      week_number:
        type: integer
    required:
    - academic_day_of_week
    - academic_year
    - calendar_date
    - semester_type
    - week_number
    type: object
  dto.CreateBuildingRequest:
    properties:
      building_name:
        type: string
      building_number:
        type: string
      campus_id:
        type: integer
    required:
    - building_name
    - building_number
    - campus_id
    type: object
  dto.CreateCampusRequest:
    properties:
      campus_name:
        type: string
      campus_short:
        type: string
    required:
    - campus_name
    - campus_short
    type: object
  dto.CreateCourseInstructorRequest:
    properties:
      class_type:
        type: string
      course:
        type: integer
      employee:
        type: integer
      hours:
        type: integer
    required:
    - class_type
    - course
    - employee
    type: object
  dto.CreateCourseRequest:
    properties:
      course_code:
        type: integer
      course_coordinator:
        type: integer
      course_language:
        type: string
      course_name:
        type: string
      ects_points:
        type: integer
      leading_unit:
        type: integer
    required:
    - course_code
    - course_coordinator
    - course_language
    - course_name
    - ects_points
    - leading_unit
    type: object
  dto.CreateCourseTypeDetailRequest:
    properties:
      class_hours:
        type: integer
      class_type:
        type: string
      course:
        type: integer
      frequency:
        type: string
      manual_weeks:
        items:
          type: integer
        type: array
      max_group_participants_number:
        type: integer
      pc_needed:
        type: boolean
      projector_needed:
        type: boolean
      slots_per_class:
        type: integer
    required:
    - class_type
    - course
    type: object
...
  models.AcademicCalendarResponse:
    properties:
      academic_day_of_week:
        type: integer
      academic_year:
        type: string
      calendar_date:
        type: string
      description:
        type: string
      semester_type:
        type: string
      week_number:
        type: integer
    type: object
  models.Building:
    properties:
      building_name:
        type: string
      building_number:
        type: string
      campus_id:
        type: integer
      id:
        type: integer
    type: object
  models.BuildingReadResponse:
    properties:
      building_name:
        type: string
      building_number:
        type: string
      campus_id:
        type: integer
      id:
        type: integer
      rooms_number:
        type: integer
    type: object
  models.Campus:
    properties:
      campus_name:
        type: string
      campus_short:
        type: string
      id:
        type: integer
    type: object
  models.Course:
    properties:
      course_code:
        type: integer
      course_coordinator:
        type: integer
      course_language:
        type: string
      course_name:
        type: string
      ects_points:
        type: integer
      leading_unit:
        type: integer
    type: object
  models.CourseTypeDetail:
    properties:
      class_hours:
        type: integer
      class_type:
        type: string
      course:
        type: integer
      frequency:
        type: string
      manual_weeks:
        items:
          type: integer
        type: array
      max_group_participants_number:
        type: integer
      pc_needed:
        type: boolean
      projector_needed:
        type: boolean
      slots_per_class:
        type: integer
    type: object
  models.CoursesInstructors:
    properties:
      class_type:
        type: string
      course:
        type: integer
      employee:
        type: integer
      hours:
        type: integer
    type: object
...
  models.PaginatedAcademicCalendarResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.AcademicCalendarResponse'
        type: array
    type: object
  models.PaginatedBuildingsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.BuildingReadResponse'
        type: array
    type: object
  models.PaginatedCampusesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Campus'
        type: array
    type: object
  models.PaginatedCourseTypeDetailsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CourseTypeDetail'
        type: array
    type: object
  models.PaginatedCoursesInstructorsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CoursesInstructors'
        type: array
    type: object
  models.PaginatedCoursesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Course'
        type: array
    type: object
...
info:
  contact: {}
paths:
  /api/v1/academic-calendar:
    get:
      description: Returns all academic calendar entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedAcademicCalendarResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get academic calendar
      tags:
      - Academic Calendar
    post:
      consumes:
      - application/json
      description: Creates a new academic calendar record.
      parameters:
      - description: Academic calendar data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAcademicCalendarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AcademicCalendarResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create academic calendar entry
      tags:
      - Academic Calendar
  /api/v1/buildings:
    get:
      description: Returns list of buildings with number of rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedBuildingsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all buildings
      tags:
      - buildings
    post:
      consumes:
      - application/json
      description: Creates a new building.
      parameters:
      - description: Building data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Building'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create building
      tags:
      - buildings
  /api/v1/campuses:
    get:
      description: Returns list of all campuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedCampusesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all campuses
      tags:
      - campuses
    post:
      consumes:
      - application/json
      description: Creates a new campus.
      parameters:
      - description: Campus data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCampusRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Campus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create campus
      tags:
      - campuses
  /api/v1/course-type-details:
    get:
      description: Returns list of course type details sorted by course and class
        type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedCourseTypeDetailsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get course type details
      tags:
      - course-type-details
    post:
      consumes:
      - application/json
      description: Creates a new course type detail entry.
      parameters:
      - description: Course type detail data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourseTypeDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CourseTypeDetail'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create course type detail
      tags:
      - course-type-details
  /api/v1/courses:
    get:
      description: Returns list of all courses sorted by course code
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedCoursesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Creates a new course.
      parameters:
      - description: Course data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Course'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create course
      tags:
      - courses
  /api/v1/courses-instructors:
    get:
      description: Returns list of course-instructor assignments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedCoursesInstructorsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get course instructors
      tags:
      - courses-instructors
    post:
      consumes:
      - application/json
      description: Assigns an employee to a course and class type (after validating
        course type existence)
      parameters:
      - description: Instructor assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourseInstructorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CoursesInstructors'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Assign instructor to course
      tags:
      - courses-instructors
...
swagger: "2.0"